fix: load AWS config and assume role
